ComputerVision/ImageClassification/FlowersRecognition/PytorchSolution/requirements.txt

pytorch==1.0.2
Pillow==7.2.0
numpy==1.19.5
matplotlib==3.4.1